pyproject.toml
[tool.poetry]
name = "homelab"
version = "0.1.0"
description = ""
authors = ["xyhhx <xyhhx@disr.it>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
bs4 = "^0.0.1"
requests = "^2.28.2"
python-dotenv = "^1.0.0"
argparse = "^1.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
